+++ adventure/src/cli-client/utils.ts
@@ -1,4 +1,15 @@
+import { HexColor } from "../dsl/hexColor";
+
 export const cls = () => process.stdout.write("\x1Bc");
+export const setColor = (color: HexColor) => {
+  const red = parseInt(color.slice(0, 2), 16);
+  const green = parseInt(color.slice(2, 4), 16);
+  const blue = parseInt(color.slice(4, 6), 16);
+  process.stdout.write(`\x1b[38;2;${red};${green};${blue}m`);
+};
+
+export const resetColor = () => process.stdout.write("\033[0m");
+
 export const exitGame = (code = 0) => {
   process.exit(code);
 };
