+++ adventure/src/index.ts
@@ -1,2 +1,3 @@
 export { world, convertGame } from "./dsl/dsl";
 export { runGame as runCLIGame } from "./cli-client/run";
+export { hexColor } from "./dsl/hexColor";
