+++ koerier/src/runText.ts
@@ -2,4 +2,4 @@ import "./locaties";
 import { convertGame, runCLIGame } from "point-n-click";
 import { Game } from "./game";
 
-convertGame<Game>(runCLIGame({ color: false }));
+convertGame<Game>(runCLIGame({ color: true }));
